requirements.txt
python-redmine
pyyaml
requests[security]
